services/agent-service/requirements.txt
fastapi[standard]==0.116.0
slowapi==0.1.9
aiohttp==3.12.14
pydantic==2.11.7

python-dotenv
pydantic-settings

langchain-core

langchain-ollama
langchain-openai
langchain-anthropic
langchain-groq
langchain-huggingface

langchain-mcp-adapters
langgraph
langgraph-checkpoint-postgres

valkey==6.1.0

pytest==8.4.1
pytest-asyncio==1.0.0